api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis[hiredis]==5.0.6
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiohttp==3.9.5
apscheduler==3.10.4
bcrypt==4.1.3
